Move model classes into models folder
