+++ pyaaas/models/anonymize_result.py
@@ -1,8 +1,8 @@
 import copy
-from collections.abc import Mapping
 
-from pyaaas.dataset import Dataset
-from pyaaas.risk_profile import RiskProfile
+from models.anonymization_metrics import AnonymizationMetrics
+from models.dataset import Dataset
+from models.risk_profile import RiskProfile
 
 
 class AnonymizeResult:
