+++ tests/pyaaas/data_generator.py
@@ -1,8 +1,8 @@
 # module for collection of test data generating functions and classes
-from pyaaas.anonymize_result import AnonymizeResult
-from pyaaas.attribute_type import AttributeType
-from pyaaas.dataset import Dataset
-from pyaaas.risk_profile import RiskProfile
+from models.anonymize_result import AnonymizeResult
+from models.attribute_type import AttributeType
+from models.dataset import Dataset
+from models.risk_profile import RiskProfile
 
 
 def id_name_dataset() -> Dataset:
