+++ tests/pyaaas/models/test_Attribute.py
@@ -1,7 +1,7 @@
 import unittest
 
-from pyaaas.attribute_type import AttributeType
-from pyaaas.dataset import Dataset
+from models.attribute_type import AttributeType
+from models.dataset import Dataset
 
 
 class AttributeTest(unittest.TestCase):
